test/docflow/backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Celery + Redis
celery>=5.3.0
redis>=5.0.0
boto3>=1.34.0

# PaddleOCR (optional - comment out if not using GPU)
python -m pip install paddlepaddle-gpu==3.2.2 -i https://www.paddlepaddle.org.cn/packages/stable/cu129/
python -m pip install paddleocr

# Utils
pillow
numpy
openpyxl>=3.1.0

# PDF and Word document processing
pypdf2>=3.0.0
python-docx>=1.1.0
pdf2image>=1.16.0

# RAG / Semantic Search (LangChain)
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface>=0.1.0
pgvector>=0.2.0

# Local LLM (Qwen2.5-VL)
transformers>=4.45.0
torch>=2.0.0
accelerate>=0.25.0
qwen-vl-utils>=0.0.8
